fix(v3): proper bf16 mixed precision via DeepSpeed (fp32 master)
Previously: model.to(bfloat16) + ds_config.bf16.enabled=false
  → params/grads/optimizer ALL bf16, no fp32 master copy, numerical drift at 1B+

Fixed: keep model fp32, ds_config.bf16.enabled=true
  → DS manages bf16 compute + fp32 master + fp32 Adam state

Impact:
- Training stability / gradient precision on long runs
- +5GB fp32 master per rank (ZeRO-2 shards opt but not params)

+++ scripts/test_muon_ds.py
@@ -60,7 +60,9 @@ def main():
             D=128, N=2, K=6, num_layers=4, D_ff=256,
             vocab_size=256, memory_layer_interval=2,
         )
-    m = NeuronSparkForCausalLM(cfg).cuda().to(torch.bfloat16)
+    # Keep params fp32; DS bf16.enabled=true manages mixed precision properly
+    # (bf16 compute + fp32 master copy + fp32 optimizer state).
+    m = NeuronSparkForCausalLM(cfg).cuda()
     promote_neuron_params_fp32(m)
 
     if args.optimizer == "muon":
